added viewmodel to info ui

+++ app/src/main/java/com/androsh/shopee/data/response/ProductDataResponse.kt
@@ -7,7 +7,7 @@ data class ProductDataResponse(
     @SerializedName("id") val id: Int,
     @SerializedName("title") val title: String,
     @SerializedName("description") val description: String,
-    @SerializedName("price") val price: Long,
+    @SerializedName("price") val price: Double,
     @SerializedName("discountPercentage") val discountPercentage: Double,
     @SerializedName("rating") val rating: Double,
     @SerializedName("stock") val stock: Int,
